SystemAbstract::dirichlet is now constraint -- more general
git-svn-id: https://geuz.org/svn/gmsh/trunk@17335 59c8cabf-10d6-4a28-b43f-3936b85956c4

+++ projects/small_fem/simulation/Test.cpp
@@ -59,7 +59,7 @@ int main(int argc, char** argv){
 
   solver.solve(a, b, x);
 
-  for(size_t i = 0; i < x.size(); i++)
+  for(int i = 0; i < x.size(); i++)
     cout << x(i) << endl;
 
   a.writeToMatlabFile("a.m", "a");
